chore: update payment QR handling and schema adjustments
- Bumped package version to 3.15.2.
- Refactored payment QR functions to handle dollar amounts instead of cents, improving clarity and consistency in payment processing.
- Updated SQL schema to change payment_qr_amount from INTEGER to NUMERIC(12, 2) for better precision.
- Adjusted various components and services to accommodate the new payment handling logic, ensuring proper conversion and formatting of dollar amounts.
- Enhanced tests to validate new payment QR functionalities and ensure regression coverage.

+++ lib/supabase/fsm-schema.sql
@@ -429,7 +429,7 @@ CREATE TABLE jobs (
     sap_cm_status VARCHAR(20),
     sap_job_income NUMERIC(14, 4) NOT NULL DEFAULT 0,
     payment_qr_uen VARCHAR(50),
-    payment_qr_amount INTEGER,
+    payment_qr_amount NUMERIC(12, 2),
     payment_qr_editable BOOLEAN DEFAULT false,
     payment_qr_expiry VARCHAR(8),
     payment_qr_ref_number VARCHAR(255),
